Consolidate daSldMng_c into one genuine TU, src/actors/d_a_sld_mng.cpp
ov019/daSldMng_c's five functions lived in five one-function files plus a
never-enrolled src_tu/ shadow that suppressed the real header and carried its
own TU-local dActor_c. They are now one enrolled translation unit under the
snake_case stem tools/tu_names.py derives from the class (daSldMng_c ->
d_a_sld_mng), covering .text 0x0211261c..0x0211277c as a single `complete`
delinks span.

WHAT THE PROMOTION BUYS, beyond one file instead of six: the TU defines
InitResources, which -- the destructor being inline in the header -- is this
class's first DECLARED non-inline virtual and therefore its KEY FUNCTION. So
mwccarm emits the whole _ZTV/_ZTI/_ZTS group here, and the entry licenses all
nine emitted RTTI records as deadstrip-data with their cartridge addresses.
tools/romdata_check.py now word-compares them, relocations applied:

  _ZTV10daSldMng_c  ov019 0x021133cc  VERIFIED, 124 bytes
  _ZTI10daSldMng_c  ov019 0x0211338c  VERIFIED,  12 bytes
  _ZTS10daSldMng_c  ov019 0x02113398  PARTIAL (the usual typeinfo-name extent)

plus the inherited fBase_c/dBase_c/dActor_c records. Under the coined name
`IceSlideManager` (renamed in the previous commit) none of that was possible --
those records reached no symbol home at all.

The two forcing scaffolds the per-function destructor sources needed --
daSldMng_c_EmitDestructor and daSldMng_c_EmitDeletingDestructor -- are gone with
them; a key-function TU materializes the inline D1/D0 pair on its own, in
cartridge order and with no homeless D2. The bodies are the recovered ones, not
the generator's shadow rewrite: DistToCPlayer(), KillAndTrackInDeathTable() and
++mState are spelled as real members, and only Sound::PlaySub keeps its mangled
declaration, for the by-value-Fix12 ABI reason stated at the call.

Also corrected two statements the promotion falsifies: include/daSldMng_c.h's
comment still described the coined spelling as a "compatibility" name and
credited the deleted scaffolds for the destructor pair, and the daSldMng_c
section of notes/player-provenance.md still said the class mirrors dActor_c
flat rather than deriving from it, and pointed at the deleted shadow.

Verification:
  tools/rombuild.py -j16   106/106 exact, 100.000000%, mismatching 0
  ROM data from source     569 verified (flat -- the deleted per-function
                           sources included the same header, so the three
                           daSldMng_c records were already in the set)
  tools/tiers_ratchet.py --check   PASS 2591/2591 (the three CONVERTED rows are
                           renamed to path#symbol, not removed)
  eligible 11096/11211, check_dead_references, layout_check, port_refcheck,
  check_src_tu_compiles 98/98, check_header_offsets -- all clean
