+++ attribution.json
@@ -1104,6 +1104,11 @@
     "src/actors/d_a_obj_waterfall.cpp#_ZN16daObjWaterfall_cD0Ev": "andrewboudreau",
     "src/actors/d_a_obj_waterfall.cpp#_ZN16daObjWaterfall_c8BehaviorEv": "github-actions[bot]",
     "src/actors/d_a_obj_waterfall.cpp#_ZN16daObjWaterfall_c13InitResourcesEv": "github-actions[bot]",
-    "src/actors/d_a_obj_waterfall.cpp#WaterfallMist_Spawn": "github-actions[bot]"
+    "src/actors/d_a_obj_waterfall.cpp#WaterfallMist_Spawn": "github-actions[bot]",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_cD1Ev": "andrewboudreau",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_cD0Ev": "andrewboudreau",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_sld_mng.cpp#daSldMng_c_Spawn": "github-actions[bot]"
   }
 }
