+++ config/arm9/overlays/ov019/delinks.txt
@@ -124,25 +124,9 @@ src/RacingPenguin_Spawn.c:
     complete
     .text start:0x021125bc end:0x0211261c
 
-src/_ZN10daSldMng_cD1Ev.cpp:
+src/actors/d_a_sld_mng.cpp:
     complete
-    .text start:0x0211261c end:0x02112640
-
-src/_ZN10daSldMng_cD0Ev.cpp:
-    complete
-    .text start:0x02112640 end:0x02112678
-
-src/_ZN10daSldMng_c8BehaviorEv.cpp:
-    complete
-    .text start:0x02112678 end:0x0211271c
-
-src/_ZN10daSldMng_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x0211271c end:0x0211274c
-
-src/daSldMng_c_Spawn.c:
-    complete
-    .text start:0x0211274c end:0x0211277c
+    .text start:0x0211261c end:0x0211277c
 
 src/__sinit_ov019_021127a4.c:
     complete
