+++ config/converted-baseline.json
@@ -890,9 +890,9 @@
     "src/_ZN15FireSeaElevator6RenderEv.cpp",
     "src/_ZN15FireSeaElevatorD0Ev.cpp",
     "src/_ZN15FireSeaElevatorD1Ev.cpp",
-    "src/_ZN10daSldMng_c13InitResourcesEv.cpp",
-    "src/_ZN10daSldMng_cD0Ev.cpp",
-    "src/_ZN10daSldMng_cD1Ev.cpp",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_c13InitResourcesEv",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_cD0Ev",
+    "src/actors/d_a_sld_mng.cpp#_ZN10daSldMng_cD1Ev",
     "src/_ZN15MaterialChanger6UpdateER15ModelComponents.cpp",
     "src/_ZN15MaterialChanger7PrepareER8BMD_FileR8BMA_File.cpp",
     "src/_ZN15MaterialChangerC1Ev.cpp",
