+++ config/tu_manifest.d/ov019/daSldMng_c.json
@@ -1,9 +1,9 @@
 {
   "id": "ov019/daSldMng_c",
   "module": "ov019",
-  "source": "src_tu/actors/daSldMng_c.cpp",
-  "promoted_source": "src/actors/daSldMng_c.cpp",
-  "status": "text-verified",
+  "source": "src/actors/d_a_sld_mng.cpp",
+  "promoted_source": "src/actors/d_a_sld_mng.cpp",
+  "status": "promoted",
   "boundary_confidence": "high",
   "boundary_evidence": [
     "contiguous linker run: 0x211261c..0x211277c, 5 function(s), from build/tu_map.json (tools/tu_map.py)",
@@ -75,5 +75,70 @@
       "functions_occur_in_expected_order": "PARTIAL -- ordinal pair(s) not in ROM order: [(0, 1)]",
       "relocation_destinations_verified": "PASS"
     }
-  }
+  },
+  "compiler_only_output": [
+    {
+      "symbol": "_ZTI7fBase_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x02086d70, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x02086d70"
+    },
+    {
+      "symbol": "_ZTS7dBase_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x02086e54, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x02086e54"
+    },
+    {
+      "symbol": "_ZTS7fBase_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x02086e60, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x02086e60"
+    },
+    {
+      "symbol": "_ZTS8dActor_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x0208e384, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x0208e384"
+    },
+    {
+      "symbol": "_ZTI7dBase_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x02086e78, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x02086e78"
+    },
+    {
+      "symbol": "_ZTI10daSldMng_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at ov019:0x0211338c, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "ov019",
+      "canonical_address": "0x0211338c"
+    },
+    {
+      "symbol": "_ZTI8dActor_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at arm9:0x0208e390, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "arm9",
+      "canonical_address": "0x0208e390"
+    },
+    {
+      "symbol": "_ZTS10daSldMng_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at ov019:0x02113398, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "ov019",
+      "canonical_address": "0x02113398"
+    },
+    {
+      "symbol": "_ZTV10daSldMng_c",
+      "disposition": "deadstrip-data",
+      "reason": "Emitted as vague linkage because this TU defines the class's key function. The cartridge keeps its canonical copy at ov019:0x021133cc, outside this entry's only licensed range (.text 0x0211261c-0x0211277c), so dsd supplies that ROM data independently. romdata_check compares the emitted copy against the cartridge, relocations applied, before production isolation discards it.",
+      "canonical_module": "ov019",
+      "canonical_address": "0x021133cc"
+    }
+  ]
 }
