+++ include/daSldMng_c.h
@@ -3,11 +3,18 @@
 
 #include "dActor_c.h"
 
-/* The cartridge RTTI names this class daSldMng_c. daSldMng_c is the
- * readable compatibility spelling already carried by all four configured
- * virtual symbols. Its __si_class_type_info record points directly at
- * dActor_c at offset zero, and its vtable has exactly the same 31 slots as
- * that base. Only InitResources (slot 0), Behavior (slot 6), and the D1/D0
+/* THE CLASS NAME IS THE CARTRIDGE'S OWN. The decomp used to call this class
+ * `IceSlideManager`, a coined name. MEASURED in
+ * extracted/overlays/overlay_0019.bin (ov019 base 0x021111a0, the `.text
+ * start:` on line 1 of config/arm9/overlays/ov019/delinks.txt): the vtable
+ * object's preamble at 0x021133c4 is [offset-to-top 0, 0x0211338c], and
+ * _ZTI10daSldMng_c at 0x0211338c reads [0x0209a764, 0x02113398, 0x0208e390] --
+ * _ZTVN3abi20__si_class_type_infoE (config/arm9/symbols.txt), a typeinfo name
+ * at 0x02113398 whose bytes are the string "10daSldMng_c", and _ZTI8dActor_c.
+ *
+ * So the __si_class_type_info record points directly at dActor_c at offset
+ * zero, and the vtable at 0x021133cc has exactly the same 31 slots as that
+ * base. Only InitResources (slot 0), Behavior (slot 6), and the D1/D0
  * destructor pair (slots 16/17) are overridden.
  *
  * The factory's literal allocation size pins the class at 0xd8. dActor_c ends
@@ -21,10 +28,21 @@ struct daSldMng_c : dActor_c {
     u8 mState;            /* 0x0d6 */
     u8 pad_0d7;           /* 0x0d7 */
 
-    /* Inline is load-bearing: explicit use from the destructor sources makes
-     * mwccarm emit D1/D0 in cartridge order without a homeless D2. */
+    /* Inline is load-bearing. Written out-of-line in the TU, mwccarm emits D0
+     * BEFORE D1 -- the reverse of the cartridge's order, which objisolate
+     * refuses for the whole translation unit -- plus a third D2 with no ROM
+     * home. Defined here it emits the retail D1/D0 pair in ROM order and no D2.
+     * The two forcing scaffolds this replaces (daSldMng_c_EmitDestructor and
+     * daSldMng_c_EmitDeletingDestructor, in the deleted per-function sources)
+     * are no longer needed: src/actors/d_a_sld_mng.cpp defines InitResources,
+     * so it is the key-function TU and materializes the pair on its own. */
     virtual ~daSldMng_c() {}
 
+    /* THE KEY FUNCTION IS InitResources -- the first DECLARED non-inline
+     * virtual, not the first slot. Whichever TU defines it emits this class's
+     * _ZTV/_ZTI/_ZTS group; that is src/actors/d_a_sld_mng.cpp, which licenses
+     * all three as deadstrip-data so romdata_check word-compares them against
+     * the cartridge. */
     virtual int InitResources(); /* slot 0 */
     virtual int Behavior();      /* slot 6 */
 };
