--- src_tu/actors/daSldMng_c.cpp
@@ -1,157 +1,0 @@
-//cpp
-/* SHADOW translation unit -- generated by tools/tubuild.py create.
- * ov019/daSldMng_c  (5 function(s))
- *
- * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
- * build. It is a STARTING POINT (plan sec 7.3): local shadow declarations
- * below were carried verbatim from the legacy files, not reconciled against
- * real project headers -- that judgement call is left to a human/LLM review,
- * the way pilot #1 reconciled PoleLift::Render and ::CleanupResources by hand
- * (see notes/tu-reconstruction-pilot-report.md sec 5.2).
- *
- * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
- * emits one .text section per function, in the REVERSE of source order, so
- * the highest-address ROM function is written FIRST here. Do not reorder;
- * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
- * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
- *
- * Assembled from these legacy one-function sources (ROM address order):
- *   [0] 0x0211261c  src/_ZN10daSldMng_cD1Ev.cpp
- *   [1] 0x02112640  src/_ZN10daSldMng_cD0Ev.c
- *   [2] 0x02112678  src/_ZN10daSldMng_c8BehaviorEv.cpp
- *   [3] 0x0211271c  src/_ZN10daSldMng_c13InitResourcesEv.cpp
- *   [4] 0x0211274c  src/daSldMng_c_Spawn.c
- */
-
-/* Includes: union of the legacy files', first-seen in ROM-ascending
- * processing order. NOT verified for header ordering constraints (e.g. a
- * common.h-before-X rule) -- watch for new compile errors after this. */
-#define DASLDMNG_C_H  /* suppress the flat C header; the TU-local class below is the C++ view, same evidenced offsets */
-#include "decl_Actor.h"
-#include "decl_common.h"
-#include "daSldMng_c.h"
-#include "decl_ActorBase.h"
-
-/* Local shadow declarations carried from the legacy files verbatim.
- * NOT reconciled against real project headers -- check include/*.h for
- * each of these before compiling; a real header should usually win. */
-/* shadow struct 'dActor_c' */
-extern "C" void _ZN6Memory10DeallocateEPvP4Heap(void *, void *);
-extern "C" void *data_020a0eac;
-
-struct dActor_c {   /* shadow base: vptr at 0, evidenced fields, size 0xd4 */
-    char pad0[0x58];
-    s32 mPosX, mPosY, mPosZ, mPrevPosX, mPrevPosY, mPrevPosZ;
-    s32 mCamSpacePosX, mCamSpacePosY, mCamSpacePosZ;
-    s32 mScaleX, mScaleY, mScaleZ;
-    s16 mAngleX, mAngleY, mAngleZ, mPrevAngleX, mPrevAngleY, mPrevAngleZ;
-    s32 mHorzSpeed, mVertAccel, mTerminalVelocity;
-    u8  pad_0a4[0x4];
-    s32 mVertSpeed;
-    u8  pad_0ac[0x4];
-    u32 mFlags;
-    s32 mClipOffsetY, mClipRadius, mClipDistance, mFarDistance;
-    u8  mClipResult;
-    u8  pad_0c5[0x7];
-    s8  mAreaId;
-    u8  pad_0cd[0x1];
-    s16 mDeathTableID;
-    u8  pad_0d0[0x4];
-    virtual ~dActor_c();
-    void operator delete(void *ptr) { _ZN6Memory10DeallocateEPvP4Heap(ptr, data_020a0eac); }
-};
-struct daSldMng_c : dActor_c {
-    s16 mKillTimer;   /* 0x0d4 */
-    u8  mState;   /* 0x0d6 */
-    virtual ~daSldMng_c();
-    int Behavior();
-    int InitResources();
-};
-
-/* shadow struct 'S3' */
-struct S3 { int w0; int w1; int w2; };
-
-extern "C" {
-extern void* data_020a0eac;
-extern int _ZTV10daSldMng_c[];
-extern int _ZN8dActor_c13DistToCPlayerEv(void*);
-extern int _ZN5Sound7PlaySubEjjj5Fix12IiEb(unsigned,unsigned,unsigned,int,int);
-extern int DecIfAbove0_Short(void*);
-extern int _ZN8dActor_c24KillAndTrackInDeathTableEv(void*);
-extern struct S3 data_ov019_021135d8;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 4 -- daSldMng_c_Spawn, 0x0211274c, size 0x30 */
-/* -------------------------------------------------------------------------- */
-// @symbol daSldMng_c_Spawn
-/* recovered: globals resolved, declarations from a shared header */
-/* recovered: globals resolved */
-/* resolved: VT = _ZTV10daSldMng_c */
-extern "C" {  /* .c-derived member: C linkage for the whole block */
-int *daSldMng_c_Spawn(void)
-{
-    int *p = (int *)_ZN7fBase_cnwEj(216);
-    if (p) { _ZN8dActor_cC2Ev(p); p[0] = (int)&_ZTV10daSldMng_c[2]; /* +8: this TU defines the vtable */ }
-    return p;
-}
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 3 -- _ZN10daSldMng_c13InitResourcesEv, 0x0211271c, size 0x30 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daSldMng_c13InitResourcesEv
-/* recovered: named members + shared header, real C++ method */
-int daSldMng_c::InitResources()
-{
-    *(int*)((char*)&mPosX) = data_ov019_021135d8.w0;
-    *(int*)((char*)&mPosY) = data_ov019_021135d8.w1;
-    *(int*)((char*)&mPosZ) = data_ov019_021135d8.w2;
-    *(short*)((char*)&mKillTimer) = 0x78;
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 2 -- _ZN10daSldMng_c8BehaviorEv, 0x02112678, size 0xa4 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daSldMng_c8BehaviorEv
-/* recovered: named members + shared header, real C++ method */
-int daSldMng_c::Behavior()
-{
-  switch (mState) {
-  case 0:
-    if (_ZN8dActor_c13DistToCPlayerEv(((char*)this)) < 0x180000) {
-      unsigned char* p = (unsigned char*)(((int)((char*)this) + 0xd6));
-      _ZN5Sound7PlaySubEjjj5Fix12IiEb(0x20, 0x14, 0x7f, 0x15666, 0);
-      *p += 1;
-    }
-    break;
-  case 1:
-    if (DecIfAbove0_Short((char*)&mKillTimer) == 0) {
-      _ZN5Sound7PlaySubEjjj5Fix12IiEb(0x20, 0x7f, 0, 0x15666, 0);
-      _ZN8dActor_c24KillAndTrackInDeathTableEv(((char*)this));
-    }
-    break;
-  }
-  return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 1 -- _ZN10daSldMng_cD0Ev, 0x02112640, size 0x38 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daSldMng_cD0Ev
-/* recovered: named members + shared header, declarations from a shared header */
-/* recovered: named members + shared header */
-extern "C" {  /* .c-derived member: C linkage for the whole block */
-/* (no separate definition: the single ~daSldMng_c() below emits the D0 and D1
- * variants together -- a hand-mangled D0 next to a real destructor is the
- * known mwccarm ICE, ELFgen.c:483.) */
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 0 -- _ZN10daSldMng_cD1Ev, 0x0211261c, size 0x24 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daSldMng_cD1Ev
-daSldMng_c::~daSldMng_c()
-{
-}
